券仓确认订单

+++ twant/lib_net/src/main/java/Store.kt
@@ -4,8 +4,8 @@ data class Store(
     val contact: String,
     val id: Int,
     val industry: String,
-    val lat: Int,
-    val lnt: Int,
+    val lat: Double,
+    val lnt: Double,
     val name: String,
     val platformName: String,
     val platformStoreId: Int
